+++ integration/playwright-test/examples/failing/test-compromised.spec.ts
@@ -1,0 +1,16 @@
+import { Ensure, equals } from '@serenity-js/assertions';
+import { TestCompromisedError } from '@serenity-js/core';
+import { describe, it } from '@serenity-js/playwright-test';
+
+describe('Playwright Test reporting', () => {
+
+    describe('A scenario', () => {
+
+        it('is compromised', async ({ actorCalled }) => {
+            await actorCalled('Alice').attemptsTo(
+                Ensure.that('Hello', equals('Hola'))
+                    .otherwiseFailWith(TestCompromisedError, 'Translation DB is down, please cheer it up'),
+            );
+        });
+    });
+});