+++ integration/playwright-test/examples/screenplay/ability-discard-error.spec.ts
@@ -1,0 +1,28 @@
+import { Interaction } from '@serenity-js/core';
+import { describe, it } from '@serenity-js/playwright-test';
+
+describe('Playwright Test reporting', () => {
+
+    describe('A screenplay scenario', () => {
+
+        it(`fails when discarding an ability fails`, async ({ actorCalled }) => {
+            await actorCalled('Donald')
+                .whoCan(new CauseErrorWhenDiscarded())
+                .attemptsTo(
+                    NotDoTooMuch(),
+                );
+        });
+    });
+});
+
+const NotDoTooMuch = () => Interaction.where(`#actor doesn't do much`, () => void 0);
+
+class CauseErrorWhenDiscarded {
+    static as(actor) {
+        return actor.abilityTo(CauseErrorWhenDiscarded);
+    }
+
+    discard() {
+        return Promise.reject(new TypeError(`Some internal error in ability`));
+    }
+}