+++ packages/core/spec/io/loader/examples/module-with-factory-method.ts
@@ -1,0 +1,1 @@
+export { ExampleWithFactoryMethod } from './ExampleWithFactoryMethod';