+++ packages/core/src/stage/Stage.ts
@@ -264,7 +264,6 @@ export class Stage {
         }
 
         // Wait for the Photographer to finish taking any screenshots
-
         await this.manager.waitForAsyncOperationsToComplete();
 
         const actorsToDismiss = new Map<Actor, CorrelationId>(actors.map(actor => [actor, CorrelationId.create()]));