+++ packages/playwright-test/src/reporter/PlaywrightAttachments.ts
@@ -1,0 +1,1 @@
+export const SERENITY_JS_DOMAIN_EVENTS_ATTACHMENT_CONTENT_TYPE = 'application/vnd.serenity-js.domain-events+json';